pyproject.toml
[project]
name = "rapidocr-mcp"
version = "1.0.0"
description = "High-performance RapidOCR MCP Server"
readme = "README.md"
requires-python = ">=3.12"
license = {text = "MIT"}
authors = [{name = "RapidOCR Team", email = "rapidocr@example.com"}]
keywords = ["mcp", "ocr", "rapidocr", "model-context-protocol", "ai"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "mcp[cli]>=1.2.0",
    "rapidocr-onnxruntime>=1.4.0",
    "pillow>=10.0.0",
    "pydantic-settings>=2.0.0",
    "loguru>=0.7.0",
    "httpx>=0.27.0",
    "pydantic>=2.0.0",
    "fastapi>=0.115.0",
    "uvicorn>=0.32.0",
    "python-dotenv>=1.0.0",
    "aiofiles>=24.0.0",
    "prometheus-client>=0.21.0",
    "opentelemetry-api>=1.25.0",
    "opentelemetry-sdk>=1.25.0",
    "opentelemetry-instrumentation-fastapi>=0.46b0",
    "opentelemetry-exporter-prometheus>=0.46b0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.8.0",
    "pytest>=8.3.0",
    "pytest-asyncio>=0.25.0",
    "pytest-cov>=6.0.0",
    "httpx>=0.27.0",
    "aioresponses>=0.7.6",
]
docker = [
    "uv>=0.5.0",
]

[project.scripts]
rapidocr-mcp = "rapidocr_mcp.__main__:main"

[tool.uv]
package = true
dev-dependencies = ["ruff>=0.8.0", "pytest>=8.3.0", "pytest-asyncio>=0.25.0", "pytest-cov>=6.0.0"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 100
target-version = "py312"
lint.select = ["E", "F", "W", "I", "N", "UP", "B", "C4", "SIM"]
lint.ignore = ["E501"]

[tool.pytest.ini_options]
pythonpath = ["src"]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "-v --cov=src --cov-report=term-missing --cov-report=html"

[tool.coverage.run]
source = ["src"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
]
